chore(client,types): extract some appMusicPlayer types nor interface into types package
