+++ packages/types/src/index.ts
@@ -40,6 +40,7 @@ export type Eheh = keyof typeof Ahah;
  */
 export const ResponseToUniverse = 42;
 
+export * from './appMusicPlayerMachine';
 export * from './http';
 export * from './TemporalToServer';
 export * from './websockets';
